+++ ndk-glue/src/lib.rs
@@ -14,6 +14,11 @@ use std::ptr::NonNull;
 use std::sync::{Arc, Condvar, Mutex, RwLock, RwLockReadGuard};
 use std::thread;
 
+#[cfg(feature = "logger")]
+pub use android_logger;
+#[cfg(feature = "logger")]
+pub use log;
+
 pub use ndk_macro::main;
 
 /// `ndk-glue` macros register the reading end of an event pipe with the