+++ ndk-macro/src/parse.rs
@@ -40,10 +40,6 @@ mod logger {
         pub tag: Option<String>,
         // Filtering rules
         pub filter: Option<String>,
-        // Path to `android_logger` to override
-        pub android_logger: Option<Path>,
-        // Path to `log` crate to override
-        pub log: Option<Path>,
     }
 
     #[derive(FromMeta, PartialEq, Eq, Debug, Clone, Copy)]
@@ -113,7 +109,7 @@ mod test {
     }
 
     #[test]
-    fn overriden_ndk_glue() {
+    fn overridden_ndk_glue() {
         let attr: AttributeArgs = parse_quote! {
             ndk_glue = "my::re::exported::ndk_glue"
         };
@@ -196,38 +192,6 @@ mod test {
 
             assert_eq!(logger.level, Some(LogLevel::Warn));
             assert_eq!(logger.tag.unwrap(), "my-app");
-        }
-
-        #[test]
-        fn overriden_android_logger() {
-            let attr: AttributeArgs = parse_quote! {
-                logger(android_logger = "my::re::exported::android_logger")
-            };
-            let attr: MainAttr = FromMeta::from_list(&attr).unwrap();
-
-            let logger = attr.logger.unwrap();
-
-            let expected_path: Path = parse_quote! {
-                my::re::exported::android_logger
-            };
-
-            assert_eq!(logger.android_logger.unwrap(), expected_path);
-        }
-
-        #[test]
-        fn overriden_log_crate() {
-            let attr: AttributeArgs = parse_quote! {
-                logger(log = "my::re::exported::log")
-            };
-            let attr: MainAttr = FromMeta::from_list(&attr).unwrap();
-
-            let logger = attr.logger.unwrap();
-
-            let expected_path: Path = parse_quote! {
-                my::re::exported::log
-            };
-
-            assert_eq!(logger.log.unwrap(), expected_path);
         }
     }
 }