+++ ndk-glue/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 0.3.1 (2021-07-29)
+
 - Looper is now created before returning from `ANativeActivity_onCreate`, solving
   race conditions in `onInputQueueCreated`.
 - Event pipe and looper are now notified of removal _before_ destroying `NativeWindow`